% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{BarProp}
\alias{BarProp}
\title{Create a proportional bar plot}
\usage{
BarProp(seurat, x_axis, split_by)
}
\arguments{
\item{x_axis}{The column in the Seurat object meta.data to use as the x axis}

\item{split_by}{The column in the Seurat object meta.data used to split each bar}

\item{object}{The Seurat object}
}
\description{
Create a proportional bar plot
}
\examples{
BarProp(seurat_obj, 'seurat_clusters', 'cell_type')
}
