% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{CleanDimPlot}
\alias{CleanDimPlot}
\title{Create a clean dimplot}
\usage{
CleanDimPlot(object, group_by)
}
\arguments{
\item{object}{The Seurat object}

\item{group_by}{The column in the Seurat object meta.data to use to group cells (cluster, sample, etc.)}
}
\description{
Create a clean dimplot
}
\examples{
CleanDimPlot(seurat_obj, 'seurat_clusters')
}
