% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ClusterProp}
\alias{ClusterProp}
\title{Create a viz with a UMAP and a bar plot showing percentage of cells assigned to clusters}
\usage{
ClusterProp(seurat, cluster_col, condition_col, replicate_col, data = FALSE)
}
\arguments{
\item{cluster_col}{The column in the Seurat object meta.data to use to determine clusters (e.g. seurat_clusters)}

\item{condition_col}{The column in the Seurat object denoting condition}

\item{replicate_col}{The column in the Seurat object denoting replicates (male, female, etc.) or cell types}

\item{data}{If TRUE, returns a dataframe}

\item{object}{The Seurat object}
}
\description{
Create a viz with a UMAP and a bar plot showing percentage of cells assigned to clusters
}
\examples{
ClusterProp(seurat_obj, cluster_col = 'seurat_clusters', condition_col = 'activation', replicate_col = 'tcell_type' = 4)
}
