% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_effects.R
\name{CompareBatchEffects}
\alias{CompareBatchEffects}
\title{Assess batch correction by Harmony by printing out side by side dim plots before and after correction}
\usage{
CompareBatchEffects(object, batch_col)
}
\arguments{
\item{object}{The Seurat object}

\item{replicate_col}{The column in the Seurat object that distinguishes replicates (eg. "patient")}
}
\value{
plot
}
\description{
Assess batch correction by Harmony by printing out side by side dim plots before and after correction
}
\examples{
CompareBatchEffects(seurat_obj, 'patient')
}
