% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scripts.R
\name{CreateCRscript_Count}
\alias{CreateCRscript_Count}
\title{Create a proportional bar plot}
\usage{
CreateCRscript_Count(
  file = NULL,
  cluster_base_directory,
  id_prefix,
  sample_prefix,
  sample_numbers,
  expect_cells,
  transcriptome
)
}
\arguments{
\item{object}{The Seurat object}

\item{x_axis}{The column in the Seurat object meta.data to use as the x axis}

\item{split_by}{The column in the Seurat object meta.data used to split each bar}
}
\description{
Create a proportional bar plot
}
\examples{
BarProp(seurat_obj, 'seurat_clusters', 'cell_type')
}
