% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{CreateLabelFile}
\alias{CreateLabelFile}
\title{Assess batch correction by Harmony by printing out side by side dim plots before and after correction}
\usage{
CreateLabelFile(
  seurat,
  cluster_col = "seurat_clusters",
  cell_type_col = "sctype_id"
)
}
\arguments{
\item{seurat}{The Seurat object}

\item{cluster_col}{Clusters in seurat object to use, default is 'seurat_clusters'}

\item{cell_type_col}{Cell type column in seurat, default is "sctype_id"}
}
\value{
data.frame
}
\description{
Assess batch correction by Harmony by printing out side by side dim plots before and after correction
}
\examples{
CreateLabelFile(seurat)
}
