% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_qc.R
\name{FilterVis_nUMI}
\alias{FilterVis_nUMI}
\title{Visualize the amount of cells removed due to nUMI filtering}
\usage{
FilterVis_nUMI(seurat, umi_thresh, group_col, umi_col = "nUMI")
}
\arguments{
\item{umi_thresh}{The label (celltype, cluster, etc) column for organizing groups}

\item{group_col}{The condition column (treatment, etc) used for comparing between}

\item{umi_col}{The background group for use in the tests (usually "control" or similar)}

\item{seurat_obj}{The Seurat object}
}
\value{
ggplot
}
\description{
Visualize the amount of cells removed due to nUMI filtering
}
\examples{
FilterVis_nUMI(seurat_obj, umi_thresh = 20000, group_col = "sample", umi_col = "nUMI")
}
