% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_qc.R
\name{FilterVis_pMito}
\alias{FilterVis_pMito}
\title{Visualize the amount of cells removed due to pMito filtering}
\usage{
FilterVis_pMito(seurat_obj, pMito_thresh, group_col, pMito_col = "pMito")
}
\arguments{
\item{seurat_obj}{The Seurat object}

\item{pMito_thresh}{The label (celltype, cluster, etc) column for organizing groups}

\item{group_col}{The condition column (treatment, etc) used for comparing between}

\item{pMito_col}{The background group for use in the tests (usually "control" or similar)}
}
\value{
ggplot
}
\description{
Visualize the amount of cells removed due to pMito filtering
}
\examples{
FilterVis_pMito(seurat_obj, pMito_thresh = 15, group_col = "sample", pMito_col = "pMito")
}
