% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{FindAllDEGs}
\alias{FindAllDEGs}
\title{Find Markers for all celltypes + conditions using Seurat FindMarkers}
\usage{
FindAllDEGs(
  seurat_obj,
  label_col,
  condition_col,
  background_group,
  test.use = "wilcox",
  assay = "RNA",
  latent.vars = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{The Seurat object}

\item{label_col}{The label (celltype, cluster, etc) column for organizing groups}

\item{condition_col}{The condition column (treatment, etc) used for comparing between}

\item{background_group}{The background group for use in the tests (usually "control" or similar)}

\item{test.use}{The type of test to use (e.g. "wilcox", "MAST", etc)}

\item{assay}{The assay to use, should be 'RNA' or 'SCT'}

\item{latent.vars}{Any variables to regress out}
}
\value{
Table with all differentially expressed genes
}
\description{
Find Markers for all celltypes + conditions using Seurat FindMarkers
}
\examples{
FindAllDEGs(seurat_obj, label_col = "celltype", condition_group = "condition", background_group = "control", test.use = "wilcox")
}
