% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_hue_pal_list}
\alias{get_hue_pal_list}
\title{Create a list of colors for a palette}
\usage{
get_hue_pal_list(n)
}
\arguments{
\item{n}{Number of colors}
}
\value{
A named list of colors to use in plotting.
}
\description{
Create a list of colors for a palette
}
\examples{
get_hue_pal_list(12)
}
