% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{MultiDimPlot}
\alias{MultiDimPlot}
\title{Create a clean dimplot split by a specific category}
\usage{
MultiDimPlot(seurat, group_by, split_by, ncol = 2, split_order = NULL)
}
\arguments{
\item{group_by}{The column in the Seurat object meta.data to use to group cells (cluster, sample, etc.)}

\item{split_by}{The column in the Seurat object to split the DimPlot by}

\item{ncol}{Number of columns in final figure (ignored for 3 or fewer plots)}

\item{split_order}{Order that split DimPlots will appear}

\item{object}{The Seurat object}
}
\description{
Create a clean dimplot split by a specific category
}
\examples{
MultiDimPlot(seurat_obj, 'seurat_clusters', 'condition', ncol = 4)
}
