% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{QuickVolcano}
\alias{QuickVolcano}
\title{Quickly create volcano plots for each celltype}
\usage{
QuickVolcano(
  de_results,
  celltype_col,
  gene_col,
  fold_change_col,
  p_val_adj,
  test_group_name = NULL,
  save = TRUE
)
}
\arguments{
\item{de_results}{A table of differential expression results grouped by celltype}

\item{celltype_col}{The label column for organizing groups}

\item{gene_col}{The column of genes}

\item{fold_change_col}{The name of the log fold change column}

\item{save}{Save individual plots for each celltype}
}
\value{
List of ggplots
}
\description{
Quickly create volcano plots for each celltype
}
\examples{
QuickVolcano(de_results, celltype_col = "celltype", gene_col = "gene", fold_change_col = "avg_log2FC", save = TRUE)
}
