% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{RunSCtype}
\alias{RunSCtype}
\title{Run sc Type on a seurat object}
\usage{
RunSCtype(
  seurat,
  tissue,
  idents = NULL,
  assay = NULL,
  new_col_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat}{The Seurat object}

\item{tissue}{The tissue within sc-type to use}

\item{idents}{The clustering column to use as input for sc-type}

\item{assay}{The assay to use, should be 'RNA' or 'SCT'}

\item{new_col_name}{The name of the new column in the seurat object with sc-type predictions}
}
\value{
Seurat object
}
\description{
Run sc Type on a seurat object
}
